Overhaul Swift package structure
