+++ StandardCyborgFusion/Sources/include/StandardCyborgFusion/SCFusionBundle.h
@@ -1,0 +1,12 @@
+//
+//  SCFusionBundle.h
+//  StandardCyborgFusion
+//
+
+#import <Foundation/Foundation.h>
+
+@interface SCFusionBundle : NSObject
+
++ (NSBundle *)fusionBundle;
+
+@end
